Move functions to own module
