+++ imod/typing/__init__.py
@@ -4,6 +4,7 @@ Module to define type aliases.
 
 from typing import TypeAlias, Union
 
+import numpy as np
 import xarray as xr
 import xugrid as xu
 
@@ -11,3 +12,5 @@ GridDataArray: TypeAlias = Union[xr.DataArray, xu.UgridDataArray]
 GridDataset: TypeAlias = Union[xr.Dataset, xu.UgridDataset]
 ScalarDataset: TypeAlias = Union[xr.Dataset, xu.UgridDataset]
 UnstructuredData: TypeAlias = Union[xu.UgridDataset, xu.UgridDataArray]
+FloatArray: TypeAlias = np.ndarray
+IntArray: TypeAlias = np.ndarray
